+++ app/tsconfig.json
@@ -16,7 +16,7 @@
     "noEmit": true,
     "jsx": "react"
   },
-  "include": ["src/**/*", "src/__tests__", "src/jest.config.ts"],
+  "include": ["src/**/*", "__tests__", "src/jest.config.ts"],
   "exclude": ["node_modules", "**/*.test.ts", "**/*.test.tsx"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
