+++ app/tsconfig.node.json
@@ -17,5 +17,5 @@
     "jsx": "react-jsx",
     "lib": ["es2017", "dom"]
   },
-  "include": ["src", "src/__tests__", "src/jest.config.ts"]
+  "include": ["src", "../__tests__", "src/jest.config.ts", "__tests__"]
 }
